Update post api validation and testing

+++ pages/api/dashboard/posts/[id].js
@@ -1,3 +1,4 @@
+import { ajv } from "@schema/validation";
 import {
   checkPerson,
   deletePost,
@@ -14,22 +15,12 @@ export default async function handler(req, res) {
       case "GET":
         const getName = req.query.name;
         const id = req.query.id;
-        // console.log(getName);
-        if (session.user.name && session.user.name == getName) {
-          // try get request, if successful return response, otherwise return error message
-          try {
-            const post = await getPostById(id);
-
-            return res.status(200).json(post);
-          } catch (err) {
-            console.error(err);
-
-            res.status(500).json({ msg: "Something went wrong." });
-          }
-        } else if (!session.user.name) {
-          const person = await checkPerson(getName);
-
-          if (person && person.email == session.user.email) {
+        if (
+          typeof getName == "string" &&
+          typeof id == "string" &&
+          id.length == 24
+        ) {
+          if (session.user.name && session.user.name == getName) {
             // try get request, if successful return response, otherwise return error message
             try {
               const post = await getPostById(id);
@@ -40,33 +31,38 @@ export default async function handler(req, res) {
 
               res.status(500).json({ msg: "Something went wrong." });
             }
+          } else if (!session.user.name) {
+            const person = await checkPerson(getName);
+
+            if (person && person.email == session.user.email) {
+              // try get request, if successful return response, otherwise return error message
+              try {
+                const post = await getPostById(id);
+
+                return res.status(200).json(post);
+              } catch (err) {
+                console.error(err);
+
+                res.status(500).json({ msg: "Something went wrong." });
+              }
+            } else {
+              res.status(401);
+            }
           } else {
             res.status(401);
           }
         } else {
-          res.status(401);
+          res.status(403);
         }
+        // console.log(getName);
 
         break;
       case "PUT":
-        const { _id, name, ...data } = req.body;
-        // console.log(req.body);
-
-        if (session.user.name && session.user.name == name) {
-          try {
-            const update = await updatePost(_id, data);
-
-            // console.log(update);
-            return res.status(200).json(update);
-          } catch (err) {
-            console.error(err);
-            res.status(500).json({ msg: "Something went wrong." });
-          }
-        } else if (!session.user.name) {
-          const person = await checkPerson(name);
-
-          if (person && person.email == session.user.email) {
-            // try get request, if successful return response, otherwise return error message
+        const { _id, ...data } = req.body;
+        const validate = ajv.getSchema("post");
+        const valid = validate(data);
+        if (typeof _id == "string" && _id.length == 24 && valid) {
+          if (session.user.name && session.user.name == data.name) {
             try {
               const update = await updatePost(_id, data);
 
@@ -76,11 +72,29 @@ export default async function handler(req, res) {
               console.error(err);
               res.status(500).json({ msg: "Something went wrong." });
             }
+          } else if (!session.user.name) {
+            const person = await checkPerson(data.name);
+
+            if (person && person.email == session.user.email) {
+              // try get request, if successful return response, otherwise return error message
+              try {
+                const update = await updatePost(_id, data);
+
+                // console.log(update);
+                return res.status(200).json(update);
+              } catch (err) {
+                console.error(err);
+                res.status(500).json({ msg: "Something went wrong." });
+              }
+            } else {
+              res.status(401);
+            }
           } else {
             res.status(401);
           }
         } else {
-          res.status(401);
+          // console.log(validate.errors);
+          res.status(403);
         }
 
         break;
@@ -92,19 +106,12 @@ export default async function handler(req, res) {
         // try delete request, if successful return response, otherwise return error message
 
         // console.log(req.body);
-        if (session.user.name && session.user.name == deleteName) {
-          try {
-            const deleted = await deletePost(deleteId);
-            return res.status(200).json(deleted);
-          } catch (err) {
-            console.error(err);
-            res.status(500).json({ msg: "Something went wrong." });
-          }
-        } else if (!session.user.name) {
-          const person = await checkPerson(deleteName);
-
-          if (person && person.email == session.user.email) {
-            // try get request, if successful return response, otherwise return error message
+        if (
+          typeof deleteId == "string" &&
+          deleteId.length == 24 &&
+          typeof deleteName == "string"
+        ) {
+          if (session.user.name && session.user.name == deleteName) {
             try {
               const deleted = await deletePost(deleteId);
               return res.status(200).json(deleted);
@@ -112,11 +119,26 @@ export default async function handler(req, res) {
               console.error(err);
               res.status(500).json({ msg: "Something went wrong." });
             }
+          } else if (!session.user.name) {
+            const person = await checkPerson(deleteName);
+
+            if (person && person.email == session.user.email) {
+              // try get request, if successful return response, otherwise return error message
+              try {
+                const deleted = await deletePost(deleteId);
+                return res.status(200).json(deleted);
+              } catch (err) {
+                console.error(err);
+                res.status(500).json({ msg: "Something went wrong." });
+              }
+            } else {
+              res.status(401);
+            }
           } else {
             res.status(401);
           }
         } else {
-          res.status(401);
+          res.status(403);
         }
 
         break;
